# Remove un-optimized "LIMIT 1" from SQL request