@pedrobaeza 2015-08-21
:+1: 

Is there the same problem on 8.0?