@mdietrichc2c 2015-08-21
@pedrobaeza seems it's also the case in 8.0; I'll do a second PR for it.